Fix computation of radiotap_len
Under gcc, when you compute the radiotap length in get_radiotap_len, it seems that you need to cast the two bytes individually to int before left-shifting or the bytes will overflow and the result will be incorrect. This pr fixes that by adding two integer casts to get_radiotap_len.